Make beam-index a standalone gRPC service
beam-index now runs as an independent gRPC server (default :50051) so
that scan operations are not duplicated across beam-stream replicas.

- Add proto/beam_index.proto with ScanLibrary RPC + build.rs codegen
- Add IndexConfig (confique) for database_url/host/port
- Add IndexServiceGrpc wrapping LocalIndexService; maps IndexError to
  gRPC Status codes
- Rewrite main.rs as a full tonic server: loads config, connects DB,
  builds all repos/services, serves gRPC
- Add GrpcIndexService to beam-stream, implementing the IndexService
  trait over a tonic Channel
- Replace LocalIndexService construction in beam-stream AppServices with
  GrpcIndexService::connect(config.beam_index_url)
- Add beam_index_url field to ServerConfig (default http://localhost:50051)
- Add tonic, prost, tonic-build workspace deps

beam-stream unit tests continue to use MockIndexService — no external
services required for cargo test --workspace.
