+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["beam-auth", "beam-stream", "beam-migration", "beam-entity"]
+members = ["beam-auth", "beam-index", "beam-stream", "beam-migration", "beam-entity"]
 resolver = "3"
 
 [workspace.dependencies]
@@ -43,6 +43,9 @@ tracing = "0.1.41"
 tracing-subscriber = "0.3.20"
 uuid = { version = "1.20.0", features = ["v4", "serde"] }
 walkdir = "2.5.0"
+tonic = { version = "0.12", features = ["transport"] }
+prost = "0.13"
+tonic-build = "0.12"
 
 [profile.release]
 codegen-units = 1
