+++ beam-index/src/models/domain/mod.rs
@@ -1,0 +1,14 @@
+pub mod admin_log;
+pub mod file;
+pub mod library;
+pub mod movie;
+pub mod show;
+pub mod stream;
+
+// Re-export all domain model types for convenience
+pub use admin_log::*;
+pub use file::*;
+pub use library::*;
+pub use movie::*;
+pub use show::*;
+pub use stream::*;
