+++ beam-index/src/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod domain;
+
+pub use domain::*;
