+++ beam-stream/Cargo.toml
@@ -9,6 +9,7 @@ default-run = "beam-stream"
 [dependencies]
 beam-auth = { path = "../beam-auth", features = ["utils"] }
 beam-entity = { path = "../beam-entity" }
+beam-index = { path = "../beam-index" }
 http = { workspace = true }
 async-graphql = { version = "7.0.17", features = [
     "apollo_persisted_queries",
@@ -38,6 +39,7 @@ m3u8-rs = "6.0.0"
 num = { version = "0.4.3", features = ["serde"] }
 num_cpus = "1.17.0"
 parking_lot = "0.12.5"
+prost = { workspace = true }
 rand = { workspace = true }
 rayon = "1.11.0"
 regex = { workspace = true }
@@ -49,6 +51,7 @@ serde_json = { workspace = true }
 tempfile = "3.23.0"
 thiserror = { workspace = true }
 tokio = { workspace = true }
+tonic = { workspace = true }
 tracing = { workspace = true, features = ["log", "release_max_level_debug"] }
 tracing-subscriber = { workspace = true, features = ["json", "env-filter"] }
 uuid = { workspace = true }
@@ -58,6 +61,7 @@ xxhash-rust = { version = "0.8.15", features = ["xxh3"] }
 [dev-dependencies]
 futures = "0.3.31"
 beam-migration = { path = "../beam-migration" }
+beam-index = { path = "../beam-index", features = ["test-utils"] }
 mockall = "0.14.0"
 
 [features]
