+++ beam-stream/src/utils/file.rs
@@ -1,7 +1,1 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum FileType {
-    Video,
-    Subtitle,
-}
+pub use beam_index::utils::file::FileType;
